+++ src/main/java/com/lgcns/backend/Like/domain/Like.java
@@ -1,0 +1,9 @@
+package com.lgcns.backend.Like.domain;
+
+import jakarta.persistence.Entity;
+import lombok.Data;
+
+@Data
+@Entity
+public class Like {
+}
