+++ src/main/java/com/lgcns/backend/Like/repository/LikeRepository.java
@@ -1,0 +1,7 @@
+package com.lgcns.backend.Like.repository;
+
+import com.lgcns.backend.Like.domain.Like;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LikeRepository extends JpaRepository<Like, Long> {
+}
