+++ src/main/java/com/lgcns/backend/comment/domain/Comment.java
@@ -1,0 +1,9 @@
+package com.lgcns.backend.comment.domain;
+
+import jakarta.persistence.Entity;
+import lombok.Data;
+
+@Data
+@Entity
+public class Comment {
+}
