+++ src/main/java/com/lgcns/backend/comment/respository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.lgcns.backend.comment.respository;
+
+import com.lgcns.backend.comment.domain.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
