+++ src/main/java/com/lgcns/backend/post/domain/Post.java
@@ -1,0 +1,10 @@
+package com.lgcns.backend.post.domain;
+
+import jakarta.persistence.Entity;
+import lombok.Data;
+
+@Entity
+@Data
+public class Post {
+
+}
